Fix the composer dependencies
It is unclear why symfony/framework-bundle was a dependency at all, it is not used in the code as is proved by maglnet/composer-require-checker.
Because only Bundle::build and Bundle::boot are used we can add a wide range of Symfony versions, from 2.8 to 5.0. The bundle is known to be used in Symfony 2.8.
Dropping support for PHP <7.2, because it is no longer supported by PHP.